+++ src/Config/Option/QrCode/DefaultBgColorConfigOption.php
@@ -8,19 +8,19 @@ use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
 use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class DefaultMarginConfigOption extends BaseConfigOption
+class DefaultBgColorConfigOption extends BaseConfigOption
 {
     public function getEnvVar(): string
     {
-        return 'DEFAULT_QR_CODE_MARGIN';
+        return 'DEFAULT_QR_CODE_BG_COLOR';
     }
 
-    public function ask(StyleInterface $io, array $currentOptions): int
+    public function ask(StyleInterface $io, array $currentOptions): ?string
     {
         return $io->ask(
-            'What\'s the default margin, in pixels, you want generated QR codes to have',
-            '0',
-            fn (mixed $value) => ConfigOptionsValidator::validateNumberGreaterThan($value, 0),
+            'What\'s the default background color for generated QR codes',
+            '#FFFFFF',
+            ConfigOptionsValidator::validateHexColor(...),
         );
     }
 }