+++ src/Config/Option/QrCode/EnabledForDisabledShortUrlsConfigOption.php
@@ -20,8 +20,6 @@ class EnabledForDisabledShortUrlsConfigOption extends BaseConfigOption
             'Should Shlink be able to generate QR codes for short URLs which are not enabled? (Short URLs are not '
             . 'enabled if they have a "valid since" in the future, a "valid until" in the past, or reached the maximum '
             . 'amount of allowed visits)',
-            // Deprecated. Shlink 4.0.0 should change default value to `true`
-            false,
         );
     }
 }