+++ src/Config/Option/Redirect/InvalidShortUrlRedirectConfigOption.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Installer\Config\Option\Redirect;
 
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
-use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class InvalidShortUrlRedirectConfigOption extends BaseConfigOption
 {
-    use ConfigOptionsValidatorsTrait;
-
     public function getEnvVar(): string
     {
         return 'DEFAULT_INVALID_SHORT_URL_REDIRECT';
@@ -23,7 +21,7 @@ class InvalidShortUrlRedirectConfigOption extends BaseConfigOption
             'Custom URL to redirect to when a user hits an invalid short URL (If no value is provided, the '
             . 'user will see a default "404 not found" page)',
             null,
-            [$this, 'validateUrl'],
+            ConfigOptionsValidator::validateUrl(...),
         );
     }
 }