+++ src/Config/Option/UrlShortener/ShortCodeLengthOption.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Installer\Config\Option\UrlShortener;
 
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
-use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class ShortCodeLengthOption extends BaseConfigOption
 {
-    use ConfigOptionsValidatorsTrait;
-
     public function getEnvVar(): string
     {
         return 'DEFAULT_SHORT_CODES_LENGTH';
@@ -23,7 +21,7 @@ class ShortCodeLengthOption extends BaseConfigOption
             'What is the default length you want generated short codes to have? (You will still be able to override '
             . 'this on every created short URL)',
             '5',
-            fn ($value) => $this->validateNumberGreaterThan($value, 4),
+            fn ($value) => ConfigOptionsValidator::validateNumberGreaterThan($value, 4),
         );
     }
 }